combine two python files into one, fix regex, make more efficient
